+++ src/components/Navigation/Navigation.css
@@ -1,10 +1,11 @@
 header {
+  overflow: hidden;
+  background-color: #fff;
   position: fixed;
   top: 0;
   left: 0;
-  width: 100vw;
-  background-color: #fff;
-  z-index: 2;
+  width: 100%;
+  z-index: 3;
 }
 
 .theme-switch {
