+++ src/components/SearchBar/SearchBar.css
@@ -5,3 +5,10 @@
   background-color: #fff;
   z-index: 1;
 }
+
+#search-container {
+  position: sticky;
+  top: 76px;
+  background-color: #fff;
+  z-index: 3;
+}
